debt: Moved onboarding components into a new module
