+++ src/drive/components/AppRoute.jsx
@@ -2,10 +2,11 @@
 import React from 'react'
 import { Route, Redirect, IndexRoute } from 'react-router'
 
+import Settings from 'drive/mobile/ducks/settings/Settings'
+import OnBoarding from 'drive/mobile/ducks/onboarding/OnBoarding'
+
 import Layout from './Layout'
 import FileExplorer from '../containers/FileExplorer'
-import Settings from '../mobile/ducks/settings/Settings'
-import OnBoarding from '../mobile/containers/OnBoarding'
 import FilesViewer from './FilesViewer'
 
 import {
