+++ src/drive/mobile/ducks/onboarding/components/Analytics.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { setAnalytics } from 'drive/mobile/ducks/settings/duck'
-import OnBoardingStep from '../../components/OnBoardingStep'
+import OnBoardingStep from './OnBoardingStep'
 
 export const Analytics = ({ onActivate, onSkip, breadcrumbs }) => (
   <OnBoardingStep
