+++ src/drive/mobile/ducks/onboarding/components/OnBoardingStep.jsx
@@ -3,7 +3,7 @@ import classNames from 'classnames'
 
 import { translate } from 'cozy-ui/react/I18n'
 import { Button } from 'cozy-ui/react'
-import styles from '../styles/onboarding'
+import styles from '../styles'
 
 export const OnBoarding = ({
   t,
