+++ src/drive/mobile/ducks/onboarding/components/Wizard.jsx
