--- src/drive/mobile/styles/breadcrumb.styl
@@ -1,18 +1,0 @@
-.onboarding-breadcrumb
-    display         flex
-    width           100%
-    margin-top      1rem
-    justify-content center
-
-    .onboarding-breadcrumb-element,
-    .onboarding-breadcrumb-element-active
-        height        .5rem
-        width         .5rem
-        border-radius 50%
-        margin        0rem .5rem
-
-    .onboarding-breadcrumb-element
-        background    #d6d8da
-
-    .onboarding-breadcrumb-element-active
-        background    #297ef2
